maint: Remove workaround to synchronize text
The issue regarding text synchronization has been resolved in LibPlay.
Thus, whenever one spanws a player with a Media.TEXT as argument, the
text promptly starts to render.

The removed workaround consisted in start a player at the very beggining
of the program with an empty text and call Player_Set_Chart () to change
the text.

* include/media/srt.ceu
* src/hello-03.ceu